test(npm): move npm download and proxy tests to test_resolver.py
Resolves https://github.com/hermetoproject/hermeto/issues/1156

Move npm dependency download and proxy behavior tests from the mixed
npm test module into tests/unit/package_managers/npm/test_resolver.py.

This keeps resolver-side network/proxy coverage grouped with resolver
logic while preserving existing assertions and mocks.

No behavior changes; only test relocation and import adjustments.
